+++ engine/src/starlight/renderer/Skybox.cpp
@@ -1,6 +1,7 @@
 #include "Skybox.hh"
 
 #include "starlight/renderer/gpu/Texture.hh"
+#include "starlight/renderer/factories/TextureFactory.hh"
 
 namespace sl {
 
@@ -11,7 +12,7 @@ Skybox::Skybox(ResourceRef<Texture> cubeMap) : m_cubeMap(cubeMap) {
 Skybox::~Skybox() { LOG_TRACE("Destroying Skybox"); }
 
 void Skybox::applyUniforms(
-  ResourceRef<Shader> shader, CommandBuffer& commandBuffer, u8 imageIndex
+  ResourceRef<Shader> shader, CommandBuffer& commandBuffer, u32 imageIndex
 ) {
     shader->setInstanceUniforms(
       commandBuffer, m_instance->getId(shader), imageIndex,
