+++ engine/src/starlight/renderer/gpu/Swapchain.hh
@@ -3,6 +3,7 @@
 #include <optional>
 
 #include "starlight/core/Core.hh"
+#include "starlight/core/math/Core.hh"
 
 #include "Sync.hh"
 #include "Texture.hh"
@@ -15,14 +16,14 @@ public:
 
     virtual void recreate(const Vec2<u32>& size) = 0;
 
-    virtual std::optional<u8> acquireNextImageIndex(
+    virtual std::optional<u32> acquireNextImageIndex(
       Semaphore* imageSemaphore = nullptr, Fence* fence = nullptr,
       Nanoseconds timeout = u64Max
-    )                                = 0;
-    virtual u8 getImageCount() const = 0;
+    )                                 = 0;
+    virtual u32 getImageCount() const = 0;
 
-    virtual Texture* getImage(u8 index) = 0;
-    virtual Texture* getDepthBuffer()   = 0;
+    virtual Texture* getImage(u32 index) = 0;
+    virtual Texture* getDepthBuffer()    = 0;
 };
 
 }  // namespace sl
