+++ engine/src/starlight/renderer/views/SkyboxRenderView.cpp
@@ -25,7 +25,7 @@ void SkyboxRenderView::init(
 void SkyboxRenderView::render(
   RendererBackend& renderer, RenderPacket& packet,
   [[maybe_unused]] const RenderProperties& properties,
-  [[maybe_unused]] float deltaTime, CommandBuffer& commandBuffer, u8 imageIndex
+  [[maybe_unused]] float deltaTime, CommandBuffer& commandBuffer, u32 imageIndex
 ) {
     auto skybox = packet.skybox;
